requirements.txt
# Core dependencies
fastapi==0.115.8
python-dotenv==1.0.1
openai==1.61.1
anthropic==0.49.0
aiohttp==3.11.12
requests==2.32.3
uvicorn==0.34.0

# Streamlit and related components
streamlit==1.42.0
streamlit-calendar==1.2.1
streamlit-authenticator==0.4.1
streamlit-antd-components==0.3.2
streamlit-autorefresh==1.0.1
streamlit-antd-components==0.3.2
streamlit-authenticator==0.4.1
google-genai==1.10.0


# Additional dependencies
phonenumbers==8.13.55
tenacity==9.0.0
apscheduler==3.10.1